アカウント登録処理とアカウント設定等の処理を分離
